Extract `annotate_model_file` and `annotate`
